Implemented field mappings
Field mappings are managed through resetFieldMappings() and
fieldMappingHandler(). Every time a PDF or CSV/TSV file is
uploaded, the fieldMappings state is reset. Every time
a mapping is updated or created, fieldMappingHandler()
first tries to delete any existing property with the
same name before adding the new mapping between
PDF field and CSV/TSV field.

VLAvailableFields now includes more of the information
that will be displayed to the user.