Pass PDF fields from the actual PDF template to the renderer.
With sufficiently large numbers of PDF form fields, this exposes
an issue with spacing, given the large cell height and lack of
scrolling for the field mapping part. The good news, obviously,
is that now this issue can be worked on effectively.

This commit also largely implements setting up pdftk for PDF
output file generation, but falls just short of making it work.
I've commented out the code that actually generates the PDFs,
so the button initiating PDF generation is currently a no-op.
I'll enable its functionality in a future commit once I've
got it working correctly.

The output file name can now be supplied as a template string,
with {@FIELD_NAME} tokens being replaced by the value for a filled
PDF form field (here: the "FIELD_NAME" field) and {#3} tokens by
the value of the given column (here: column with index 3).
Going forward, the UI will have to provide some way of setting the
file name template - right now the template string is hardcoded to
"Tax Receipt {@TR_NUMBER} - {@NAME}.pdf".