Make PDF batch generation work (with caveats)
There were a few minor bugs:
- replaceOutputPathTokens() lacked a return statement,
- the accumulator to create fieldNameReplacements was uninitialized,
- when a row was skipped, the row index was never incremented,
- fillForm() was passed filledFormFields, which wasn't accessible there.

After fixing these issues and uncommenting the pdftk invocation,
PDFs are now successfully generated.

Furthermore, this commit changes the meaning of "{#1}" column
placeholders in the output file template string to mean the
first column of the table, i.e. column number 1, instead of the
column with index 1 (which would have been the second column).
Starting the column count at 1 is more intuitive to most people
outside software engineering, but just as importent, is
consistent with numbering schemes in common spreadsheet apps.

(For consistency, the somewhat less important row index is
also swapped for row numbers, currently console output only.)

The main caveat is that the fixed text mapping, using the text
input field, doesn't work. I haven't tracked down the reason yet.
This should be easy enough to fix.

There's also no escaping or replacement of slashes/backslashes
in string substitutions for the PDF output filename template.
This could result in bad things happening. While I think that
there may be a genuine use case for parameterizing parts of
the output path including a folder name, I also believe that
row values should be escaped by default when used as
substitutions for the output path.

Another thing that should be fixed is the fact that duplicate
template string replacement values will result in output files
being overwritten, with a potential for two writes happening
at the same time and messing up the data badly. Not only that,
but also the user doesn't get told about it at any time.
For the most part, this could be mitigated before even running
batch PDF file generation.

The caveat that already existed is that the PDF output filename
template string is currently hardcoded in the renderer. We need
to integrate configuration of the filename template string into
the UI flow.

Similarly, whether or not to skip the first row (often a header)
is hardcoded as well - it's always being skipped right now.
We'll want to provide an option for this as well eventually,
although it doesn't seem too big of a deal to me compared to
the other listed shortcomings.