Bug Fix fieldMappings On Custom Text
Before the fix, the user was able to fill out custom text but
the text wasn't being sent to fieldMappings. As it turns out,
the custom text was being stored in the local state, but it
wasn't immediately updated using the handler from the parent
component. However, if the user entered custom text and then
selected it under the drop down options, fieldMappings would
have been informed.

The fix is in the submitCustomValue method. Previously, the
method only updated local state; now, onFieldMappingChange
is also called, just as it's called in mapInput.

The adjacent setState call was simplified too.