Lifting state up of display component
VLFieldMappingInput previously had self-contained state, but the
information flowed against the current of the rest of the app and
the increase in listeners was negatively affecting performance.

Lifting state centralizes information and handlers. VLFieldMappingInput
has been converted to a functional component, which reflects
its presentational-only nature. Further refactoring may be needed
to help unpacking parameters and calls to setState.