Lifting state up of display component (#7)
VLFieldMappingInput previously had self-contained state, but the
information flowed against the current of the rest of the app and
the increase in listeners was negatively affecting performance.

Lifting state centralizes information and handlers. VLFieldMappingInput
has been converted to a functional component, which reflects
its presentational-only nature.

Now, when the CSV file loads, the form fields reset because
they are controlled form fields with state centralized in the App
component.