Corrected application logic
Previously, the code mapped availableFieldsState from an
older version of pdfFields every time a new set of
pdfFields became available. Now, availableFieldsState
is current to the information immediately after a
PDF is loaded. In addition, I've removed the
conditional mounting of <VLFieldMappingInput>
as it already responds correctly to the changes in
pdfFields and csvFields. In place, I've made the CSS
display attribute responsive.