Support 'configure-system-requirements' message in the main process.
This will check the existence of the pdftk binary in the PATH,
as well as return selected OS info details. The command-exists
module and the systeminformation module are added as dependencies
to help with these tasks.

Use of 'configure-system-requirements' by the renderer is not
strictly required, but highly recommended. As a response to
sending this message, a 'system-requirements-status' message
with "status" parameter will be sent back to the renderer.
If status.isUsable is false, the renderer could disable
certain actions and instead guide the user towards installing
external dependencies, i.e. pdftk. More details are also provided.